Xoa bo thu muc scenarios
