--- KDDProject/src/test/java/scenarios/account/LoginScenarios.java
@@ -1,22 +1,0 @@
-package scenarios.account;
-
-import org.testng.annotations.Test;
-
-import domain.builder.account.LoginDetailBuilder;
-import domain.detail.account.LoginDetails;
-import scenarios.base.BaseScenario;
-
-
-//
-//public class LoginScenarios extends BaseScenario{
-//
-//    @Test
-//    public void testLoginSucessfull(){
-//        LoginDetails loginDetails = new LoginDetailBuilder().
-//        		withUsername("010203123").withPassword("123")
-//        		.build();
-//       
-////        given(user).clickLoginWith(loginDetails);
-//
-//    }
-//}
