+++ KDDProject/src/test/java/step/BaseScenario.java
@@ -1,0 +1,67 @@
+package step;
+
+import org.apache.commons.io.FileUtils;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.openqa.selenium.WebDriver;
+import org.testng.ITestResult;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import cucumber.api.java.After;
+import cucumber.api.java.Before;
+import spec.Specification;
+import util.PageStore;
+
+import java.io.File;
+import java.io.IOException;
+
+public class BaseScenario {
+	// Use the application driver
+	protected WebDriver driver;
+	public Specification user;
+	PageStore pageStore;
+
+	public BaseScenario(WebDriver driver) throws IOException {
+		this.driver = driver;
+		this.pageStore = new PageStore(driver);
+		this.user = new Specification(pageStore);
+		
+
+	}
+
+	
+	public void teardown(ITestResult result) throws IOException {
+		captureScreenShotInCaseOfFailure(result);
+		// close the browser
+		// pageStore.destroy();
+
+	}
+
+	protected void captureScreenShotInCaseOfFailure(ITestResult result) {
+		if (result.isSuccess()) {
+			return;
+		}
+		String screenShotFolder = "screenshots";
+		File screenShotSourceFile = ((TakesScreenshot) pageStore.getDriver()).getScreenshotAs(OutputType.FILE);
+		try {
+			createFolder(screenShotFolder);
+			String fileName = result.getMethod().getMethodName();
+			File screenShotTargetFile = getTargetFile(screenShotFolder, fileName, "png");
+			FileUtils.copyFile(screenShotSourceFile, screenShotTargetFile);
+		} catch (IOException e) {
+			e.printStackTrace(); 
+		}
+	}
+
+	private void createFolder(String folderName) throws IOException {
+		if (!(new File(folderName).exists()))
+			new File(folderName).mkdir();
+	}
+
+	private File getTargetFile(String folderName, String fileName, String ext) throws IOException {
+		String rootPath = new File(".").getCanonicalPath();
+		String fullPath = String.format("%s//%s//%s.%s", rootPath, folderName, fileName, ext);
+		return new File(fullPath);
+	}
+}
